+++ lib/models/primary_source.dart
@@ -1,0 +1,19 @@
+class PrimarySource {
+  final String title;
+  final String date;
+  final String content;
+  final String features;
+  final String linkTitle;
+  final String linkUrl;
+  final String imagePath;
+
+  PrimarySource({
+    required this.title,
+    required this.date,
+    required this.content,
+    required this.features,
+    required this.linkTitle,
+    required this.linkUrl,
+    required this.imagePath,
+  });
+}