+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this application will be documented here.
 - Main menu (left side);
 - Main screen (List of topics);
 - Topic screen (Markdown content)
+- Primary Sources screen
 - Settings screen (Change language)
 - About screen (Common info, Contact link, Acknowledgments and 3rd party licenses, Changelog);